rename eventlogger to events
